Poor attempts at climbing
Consider putting a timer on links to make them from one way to two way, or converting them from a temp area to another area and have invisible scouts going around laying nodes.

Alternatively, put a script on the player that leaves trails of links every half second or so.